Modularize location search
