+++ screens/LocationSearch/index.ts
@@ -1,1 +1,6 @@
-export * from "./UI"
+export * from "./Picker"
+export * from "./OptionData"
+export * from "./OptionView"
+export * from "./OptionsList"
+export * from "./UserLocationOptionView"
+export * from "./SearchBar"
