+++ screens/LocationSearch/state.ts
@@ -1,0 +1,3 @@
+import { atomWithDebounce } from "@lib/Jotai"
+
+export const searchTextAtom = atomWithDebounce("", 100)
